+++ lightsim2grid/tests/test_Iterators.py
@@ -70,9 +70,9 @@ class MakeTests(unittest.TestCase):
         assert el.bus_lv_id == self.env.backend.line_ex_to_subid[trafo_id], \
             f"trafo {trafo_id} is connected to wrong bus (lv side)"
         assert el.has_res, f"trafo {trafo_id} don't have any results"
-        assert np.abs(el.res_p_hv_mw - p_or[trafo_id]) <= tol, f"trafo {trafo_id} has wrong p_or"
-        assert np.abs(el.res_q_hv_mvar - q_or[trafo_id]) <= tol, f"trafo {trafo_id} has wrong q_or"
-        assert np.abs(el.res_a_hv_a - 0.001 * a_or[trafo_id]) <= tol, f"trafo {trafo_id} has wrong a_or"
+        assert np.abs(el.res_p_hv_mw - p_or[trafo_id]) <= tol, f"trafo {trafo_id} has wrong p_hv"
+        assert np.abs(el.res_q_hv_mvar - q_or[trafo_id]) <= tol, f"trafo {trafo_id} has wrong q_hv"
+        assert np.abs(el.res_a_hv_a - 0.001 * a_or[trafo_id]) <= tol, f"trafo {trafo_id} has wrong a_hv"
 
     def test_getters_trafo(self):
         """test that the trafo getter return the right values"""
@@ -98,6 +98,44 @@ class MakeTests(unittest.TestCase):
 
         with self.assertRaises(ValueError):
             gen_info = data_trafos[nb_trafo]
+
+    def aux_line_ok(self, el, line_id, tol, data_ref):
+        """trafo_id: id for the trafo in grid2Op"""
+        p_or, q_or, v_or, a_or = data_ref
+        assert el.connected is True, f"line {line_id} is not connected"
+        assert el.bus_or_id == self.env.backend.line_or_to_subid[line_id], \
+            f"line {line_id} is connected to wrong bus (or side)"
+        assert el.bus_ex_id == self.env.backend.line_ex_to_subid[line_id], \
+            f"line {line_id} is connected to wrong bus (ex side)"
+        assert el.has_res, f"line {line_id} don't have any results"
+        assert np.abs(el.res_p_or_mw - p_or[line_id]) <= tol, f"line {line_id} has wrong p_or"
+        assert np.abs(el.res_q_or_mvar - q_or[line_id]) <= tol, f"line {line_id} has wrong q_or"
+        assert np.abs(el.res_a_or_a - 0.001 * a_or[line_id]) <= tol, f"line {line_id} has wrong a_or"
+
+    def test_getters_line(self):
+        """test that the line getter return the right values"""
+        data_lines = self.env.backend._grid.get_lines()
+        p_or, q_or, v_or, a_or = self.env.backend.lines_or_info()
+        nb_trafo = 5
+        nb_line = 15
+        data_ref = (p_or, q_or, v_or, a_or)
+        assert len(data_lines) == nb_line
+        tol = 1e-5
+        for el in data_lines:
+            line_id = el.id
+            self.aux_line_ok(el, line_id, tol, data_ref)
+
+            with self.assertRaises(AttributeError):
+                # this should not be possible
+                el.bus_id = 2
+
+        gen_info = data_lines[0]
+        self.aux_line_ok(gen_info, 0, tol, data_ref)
+        with self.assertRaises(ValueError):
+            gen_info = data_lines[-1]
+
+        with self.assertRaises(ValueError):
+            gen_info = data_lines[nb_line]
 
 
 if __name__ == "__main__":