+++ CHANGELOG.rst
@@ -13,7 +13,11 @@ Change Log
 
 [0.5.2] 2021-xx-yy
 -------------------
+- [FIXED] `GridModel` now properly throw "out_of_range" exception when trying to change the bus of non existing
+  elements
+- [ADDED] now able to retrieve the powerlines parameters python side.
 - [IMPROVED] more explicit error messages when the building of the Ybus matrix fails.
+- [IMPROVED] now the solver is not reset when using the `backend._grid.check_solution`
 
 [0.5.1] 2021-04-09
 -------------------